split MINIT for WeekDay
