+++ package.xml
@@ -89,6 +89,8 @@ Fix: ensure build directories are created for each source directory, fixing an i
             <file name="locale_text_info.h" role="src"/>
             <file name="locale_week_day.c" role="src"/>
             <file name="locale_week_day.h" role="src"/>
+            <file name="locale_week_day_arginfo.h" role="src"/>
+            <file name="locale_week_day.stub.php" role="src"/>
             <file name="locale_week_info.c" role="src"/>
             <file name="locale_week_info.h" role="src"/>
             <file name="locale_week_info_arginfo.h" role="src"/>
