+++ src/php/ecma_intl.stub.php
@@ -710,21 +710,4 @@ namespace Ecma\Intl\Locale
         {
         }
     }
-
-    /**
-     * Days of the week.
-     *
-     * Each case is mapped to an integer value corresponding to the values
-     * defined in {@link https://tc39.es/proposal-intl-locale-info/#table-locale-weekinfo-record WeekInfo Record Fields}.
-     */
-    enum WeekDay: int
-    {
-        case Monday = 1;
-        case Tuesday = 2;
-        case Wednesday = 3;
-        case Thursday = 4;
-        case Friday = 5;
-        case Saturday = 6;
-        case Sunday = 7;
-    }
 }
